requirements.txt
Flask==2.0.1
python-telegram-bot==20.0
beautifulsoup4==4.10.0
requests==2.26.0
pymupdf==1.18.19
chromadb==0.3.21
openai==0.11.3
python-dotenv==0.19.2
lxml==4.6.3